+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## v0.5.0-beta
+
+* Patch nginx using sed instead of patch strings! Thanks to @eN0RM
+* Added a check for port 80 being used by nginx #58
+
+## v0.4.0-beta
+
+* Added ability to request certificate for more than one domain thanks to @pulecp #53 
+* Automatically get zimbraPublicServiceHostname hosts from domain and add them to the cert #54
+* Removed domain confirmation prompt for new requests. Just use `zmhostname` by default
+
+## v0.3.0-alpha
+
+* Fix *prepare certificate*, thanks to Antonio Prado (#44)
+* Added patches for Zimbra 8.8.8
+
 ## v0.2.2-alpha
 
 * Added --patch-only option to only patch nginx, to be called before renewal